- directly using enum instead of int saves alot of converting. - trigger testing

+++ src/Orchestrator/CloudRepublic.BenchMark.Orchestrator/Infrastructure/ResultConverter.cs
@@ -1,8 +1,8 @@
+using CloudRepublic.BenchMark.Application.Models;
+using CloudRepublic.BenchMark.Domain.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using CloudRepublic.BenchMark.Application.Models;
-using CloudRepublic.BenchMark.Domain.Entities;
 
 namespace CloudRepublic.BenchMark.Orchestrator.Infrastructure
 {
@@ -10,13 +10,13 @@ namespace CloudRepublic.BenchMark.Orchestrator.Infrastructure
     {
         public static IEnumerable<BenchMarkResult> ConvertToResultObject(
             IEnumerable<BenchMarkResponse> benchMarkResponses,
-            BenchMarkType benchMarkType,bool isColdRequest)
+            BenchMarkType benchMarkType, bool isColdRequest)
         {
             return benchMarkResponses.Select(benchMarkResponse => new BenchMarkResult
             {
-                CloudProvider = (int) benchMarkType.CloudProvider,
-                HostingEnvironment = (int) benchMarkType.HostEnvironment,
-                Runtime = (int) benchMarkType.Runtime,
+                CloudProvider = benchMarkType.CloudProvider,
+                HostingEnvironment = benchMarkType.HostEnvironment,
+                Runtime = benchMarkType.Runtime,
                 Success = benchMarkResponse.Success,
                 RequestDuration = Convert.ToInt32(benchMarkResponse.Duration),
                 IsColdRequest = isColdRequest
